requirements.txt
pandas
streamlit
Pillow
tqdm
torch
scikit-learn
numpy
seaborn
matplotlib